+++ src/Products/PloneMeeting/tests/testSearches.py
@@ -881,7 +881,7 @@ class testSearches(PloneMeetingTestCase):
         self.failUnless(len(collection.results()) == 1)
         self.failUnless(collection.results()[0].UID == item2.UID())
 
-    def test_pm_SearchItemsToCorrect(self):
+    def test_pm_SearchItemsToCorrectDefault(self):
         '''Test the 'items-to-correct' CompoundCriterion adapter.  This should return
            a list of items in state 'returned_to_proposing_group' the current user is able to edit.'''
         cfg = self.meetingConfig
@@ -933,8 +933,9 @@ class testSearches(PloneMeetingTestCase):
         self.do(developersItem, 'return_to_proposing_group')
         self.do(vendorsItem, 'return_to_proposing_group')
 
-        # pmManager may only edit developersItem
+        # pmManager may only manage developersItem
         self.assertTrue(self.hasPermission(ModifyPortalContent, developersItem))
+        self.assertTrue(self.hasPermission(ModifyPortalContent, vendorsItem))
         cleanRamCacheFor('Products.PloneMeeting.adapters.query_itemstocorrect')
         res = collection.results()
         self.failUnless(len(res) == 1)